refactor(frontend): split evidence-extractor.ts (832 lines) into domain modules
Extract 9 independent extractor functions into hooks/chat/evidence/ directory:
- court.ts — court decisions (search, chain, pro/contra)
- legislation.ts — legislation articles and search
- procedural-norms.ts — procedural norm text extraction
- vault.ts — document store (list, search, parse, compare)
- legal-sources.ts — retrieve_legal_sources combined handler
- rada.ts — parliament bills, deputies, voting records
- registry.ts — OpenReyestr entities, beneficiaries, enforcement, bankruptcy
- procedural.ts — deadlines, checklists, monetary claims
- dd.ts — due diligence reports, risk scoring

Shared utilities in parse.ts (content parsing, document type classification)
and types.ts (EvidenceResult interface). Barrel index.ts dispatches and merges.
Original evidence-extractor.ts reduced to a re-export shim.
