+++ lexwebapp/src/hooks/chat/evidence-extractor.ts
@@ -1,832 +1,5 @@
 /**
- * Evidence extraction from MCP tool results.
- * Transforms raw tool responses into standardized Decision[], Citation[], VaultDocument[].
+ * Re-export from evidence/ directory for backward compatibility.
  */
-
-import type { Decision, Citation, VaultDocument } from '../../types/models/Message';
-
-/**
- * Parse raw tool result content — handles MCP content array and plain objects.
- */
-function parseToolResultContent(result: any): any {
-  if (!result) return null;
-  try {
-    if (result.content && Array.isArray(result.content)) {
-      const textBlock = result.content.find((b: any) => b.type === 'text');
-      if (textBlock?.text) {
-        return JSON.parse(textBlock.text);
-      }
-    }
-    return typeof result === 'string' ? JSON.parse(result) : result;
-  } catch {
-    return result;
-  }
-}
-
-/**
- * Classify court document type from available fields (title, summary, judgment_form).
- * Used to split decisions into "Рішення" tab vs "Документи" tab in the right panel.
- */
-export function classifyDocumentType(item: any): string {
-  const form = item?.judgment_form || item?.form_name || item?.judgment_form_name
-    || item?.document_type || '';
-  const formLower = String(form).toLowerCase();
-  if (formLower.includes('постанова')) return 'Постанова';
-  if (formLower.includes('рішення')) return 'Рішення';
-  if (formLower.includes('ухвала')) return 'Ухвала';
-  if (formLower.includes('вирок')) return 'Вирок';
-  if (formLower.includes('окрема думка')) return 'Окрема думка';
-  if (formLower.includes('окрема')) return 'Окрема ухвала';
-
-  const text = [item?.title, item?.resolution, item?.summary, item?.snippet]
-    .filter(Boolean).join(' ');
-  if (/Постанова\b/i.test(text)) return 'Постанова';
-  if (/\bРішення\b/i.test(text)) return 'Рішення';
-  if (/\bУхвала\b/i.test(text)) return 'Ухвала';
-  if (/\bВирок\b/i.test(text)) return 'Вирок';
-  if (/Окрема думка/i.test(text)) return 'Окрема думка';
-
-  return '';
-}
-
-interface EvidenceResult {
-  decisions: Decision[];
-  citations: Citation[];
-  documents: VaultDocument[];
-}
-
-const courtDocUrl = (docId: string | number | undefined) =>
-  docId ? `https://reyestr.court.gov.ua/Review/${docId}` : undefined;
-
-// ---- Extractors by tool category ----
-
-function extractCourtEvidence(toolName: string, parsed: any): EvidenceResult {
-  const decisions: Decision[] = [];
-  const courtTools = [
-    'search_legal_precedents',
-    'search_supreme_court_practice',
-    'get_case_documents_chain',
-    'find_similar_fact_pattern_cases',
-    'compare_practice_pro_contra',
-    'get_court_decision',
-    'count_cases_by_party',
-  ];
-  if (!courtTools.some((t) => toolName.includes(t) || toolName === t)) {
-    return { decisions, citations: [], documents: [] };
-  }
-
-  // source_case (single)
-  if (parsed.source_case) {
-    const sc = parsed.source_case;
-    decisions.push({
-      id: `sc-${sc.doc_id || Date.now()}`,
-      number: sc.cause_num || sc.case_number || 'N/A',
-      court: sc.court_code || sc.court || '',
-      date: sc.adjudication_date || sc.date || '',
-      summary: sc.title || sc.resolution || '',
-      relevance: 100,
-      status: 'active',
-      documentType: classifyDocumentType(sc),
-      externalUrl: courtDocUrl(sc.doc_id),
-    });
-  }
-
-  // similar_cases / results array
-  const cases = parsed.similar_cases || parsed.results || parsed.cases || parsed.precedents || [];
-  for (const c of cases) {
-    decisions.push({
-      id: `d-${c.doc_id || c.id || Math.random().toString(36).slice(2, 8)}`,
-      number: c.cause_num || c.case_number || c.number || 'N/A',
-      court: c.court_code || c.court || '',
-      date: c.adjudication_date || c.date || '',
-      summary: c.title || c.resolution || c.summary || c.similarity_reason
-        || (Array.isArray(c.snippets) ? c.snippets.join(' ') : '') || '',
-      relevance: c.similarity
-        ? Math.round(c.similarity * 100)
-        : c.relevance
-          ? Math.round(c.relevance * 100)
-          : 70,
-      status: 'active',
-      documentType: classifyDocumentType(c),
-      externalUrl: courtDocUrl(c.doc_id),
-    });
-  }
-
-  // get_case_documents_chain format
-  let chainDocs: any[] = [];
-  if (parsed.documents && Array.isArray(parsed.documents)) {
-    chainDocs = parsed.documents;
-  } else if (parsed.grouped_documents && typeof parsed.grouped_documents === 'object') {
-    chainDocs = Object.values(parsed.grouped_documents).flat();
-  }
-  for (const doc of chainDocs) {
-    decisions.push({
-      id: `chain-${doc.doc_id || Math.random().toString(36).slice(2, 8)}`,
-      number: doc.case_number || parsed.case_number || doc.title || 'N/A',
-      court: doc.court || doc.instance || '',
-      date: doc.date || '',
-      summary: doc.resolution || doc.title || '',
-      relevance: 80,
-      status: 'active',
-      documentType: classifyDocumentType(doc),
-      externalUrl: courtDocUrl(doc.doc_id),
-    });
-  }
-
-  // compare_practice_pro_contra format
-  const proContraCases = [...(parsed.pro || []), ...(parsed.contra || [])];
-  for (const c of proContraCases) {
-    decisions.push({
-      id: `pc-${c.doc_id || Math.random().toString(36).slice(2, 8)}`,
-      number: c.case_number || 'N/A',
-      court: c.court || c.chamber || '',
-      date: c.date || '',
-      summary: c.snippet || '',
-      relevance: 70,
-      status: 'active',
-      documentType: classifyDocumentType(c),
-      externalUrl: courtDocUrl(c.doc_id),
-    });
-  }
-
-  // get_court_decision — single decision with sections
-  if (parsed.sections && Array.isArray(parsed.sections) && (parsed.doc_id || parsed.case_number)) {
-    const summarySection = parsed.sections.find((s: any) => s.type === 'DECISION' || s.type === 'COURT_REASONING');
-    decisions.push({
-      id: `gcd-${parsed.doc_id || Date.now()}`,
-      number: parsed.case_number || String(parsed.doc_id) || 'N/A',
-      court: '',
-      date: '',
-      summary: summarySection?.text?.slice(0, 300) || '',
-      relevance: 100,
-      status: 'active',
-      externalUrl: courtDocUrl(parsed.doc_id),
-    });
-  }
-
-  return { decisions, citations: [], documents: [] };
-}
-
-function extractLegislationEvidence(toolName: string, parsed: any): EvidenceResult {
-  const citations: Citation[] = [];
-  const legislationTools = [
-    'search_legislation',
-    'get_legislation_article',
-    'get_legislation_articles',
-    'get_legislation_section',
-    'find_relevant_law_articles',
-  ];
-  if (!legislationTools.some((t) => toolName.includes(t) || toolName === t)) {
-    return { decisions: [], citations, documents: [] };
-  }
-
-  // Single article result
-  if (parsed.full_text || parsed.text || parsed.content) {
-    const articleNum = parsed.article_number || parsed.section_name || '';
-    const title = parsed.title || parsed.rada_id || parsed.legislation_id || '';
-    citations.push({
-      text: parsed.full_text || parsed.text || parsed.content || '',
-      source: articleNum ? `${title}, ст. ${articleNum}` : title,
-    });
-  }
-
-  // Array of legislation results
-  const legislationArray = parsed.legislation || (toolName === 'search_legislation' ? parsed.articles : null);
-  if (legislationArray && Array.isArray(legislationArray)) {
-    for (const l of legislationArray) {
-      citations.push({
-        text: l.full_text || l.text || l.snippet || l.title || '',
-        source: l.article_number
-          ? `${l.title || l.rada_id || 'Норма'}, ст. ${l.article_number}`
-          : (l.title || l.type || 'Нормативний акт'),
-      });
-    }
-  }
-
-  // find_relevant_law_articles
-  if (toolName === 'find_relevant_law_articles') {
-    const refs = parsed.relevant_articles || parsed.articles || (Array.isArray(parsed) ? parsed : []);
-    for (const r of refs) {
-      if (typeof r === 'string') {
-        citations.push({ text: r, source: r });
-      } else if (r?.article || r?.reference || r?.norm) {
-        citations.push({
-          text: r.text || r.content || r.description || r.article || r.reference || r.norm || '',
-          source: r.article || r.reference || r.norm || r.title || 'Норма',
-        });
-      }
-    }
-  } else if (parsed.articles && Array.isArray(parsed.articles)) {
-    for (const a of parsed.articles) {
-      if (typeof a === 'object' && a !== null) {
-        citations.push({
-          text: a.full_text || a.text || a.content || '',
-          source: `Стаття ${a.article_number || ''}`,
-        });
-      }
-    }
-  }
-
-  return { decisions: [], citations, documents: [] };
-}
-
-function extractProceduralNormEvidence(toolName: string, rawResult: any): EvidenceResult {
-  const citations: Citation[] = [];
-  const proceduralNormTools = [
-    'search_procedural_norms',
-    'calculate_procedural_deadlines',
-    'build_procedural_checklist',
-  ];
-  if (!proceduralNormTools.some((t) => toolName === t)) {
-    return { decisions: [], citations, documents: [] };
-  }
-
-  const textContent = rawResult?.content?.find((b: any) => b.type === 'text')?.text;
-  if (textContent) {
-    const sourceLabel =
-      toolName === 'search_procedural_norms' ? 'Процесуальна норма' :
-      toolName === 'calculate_procedural_deadlines' ? 'Процесуальні строки' :
-      'Процесуальний чеклист';
-    citations.push({ text: textContent, source: sourceLabel });
-  }
-
-  return { decisions: [], citations, documents: [] };
-}
-
-function extractVaultEvidence(toolName: string, parsed: any): EvidenceResult {
-  const citations: Citation[] = [];
-  const documents: VaultDocument[] = [];
-  const vaultTools = [
-    'list_documents',
-    'semantic_search',
-    'semantic_search_vault',
-    'get_document',
-    'store_document',
-    'parse_document',
-    'extract_document_sections',
-    'summarize_document',
-    'compare_documents',
-    'extract_key_clauses',
-  ];
-  if (!vaultTools.some((t) => toolName.includes(t) || toolName === t)) {
-    return { decisions: [], citations, documents };
-  }
-
-  // list_documents
-  if (parsed.documents && Array.isArray(parsed.documents)) {
-    for (const doc of parsed.documents) {
-      documents.push({
-        id: doc.id || `vd-${Math.random().toString(36).slice(2, 8)}`,
-        title: doc.title || doc.name || 'Без назви',
-        type: doc.type || 'other',
-        uploadedAt: doc.created_at || doc.uploadedAt || doc.uploaded_at || '',
-        metadata: doc.metadata || {},
-      });
-    }
-  }
-
-  // semantic_search
-  if (Array.isArray(parsed) && parsed.length > 0 && parsed[0].documentId) {
-    for (const r of parsed) {
-      documents.push({
-        id: r.documentId || `vd-${Math.random().toString(36).slice(2, 8)}`,
-        title: r.title || r.sectionTitle || 'Без назви',
-        type: r.type || 'other',
-        metadata: { relevance: r.relevance, snippet: r.text?.slice(0, 200) },
-      });
-    }
-  }
-
-  // get_document / store_document — single
-  if (parsed.id && parsed.title && !parsed.documents) {
-    documents.push({
-      id: parsed.id,
-      title: parsed.title,
-      type: parsed.type || 'other',
-      uploadedAt: parsed.created_at || parsed.uploadedAt || '',
-      metadata: parsed.metadata || {},
-    });
-  }
-
-  // parse_document / summarize_document / extract_key_clauses
-  if (parsed.summary || parsed.text || parsed.clauses || parsed.sections) {
-    const content = parsed.summary || parsed.text || '';
-    if (content) {
-      citations.push({
-        text: content,
-        source: parsed.title || parsed.document_id || toolName,
-      });
-    }
-    if (parsed.clauses && Array.isArray(parsed.clauses)) {
-      for (const clause of parsed.clauses) {
-        citations.push({
-          text: clause.text || clause.content || '',
-          source: clause.title || clause.name || 'Ключове положення',
-        });
-      }
-    }
-    if (parsed.sections && Array.isArray(parsed.sections)) {
-      for (const sec of parsed.sections) {
-        citations.push({
-          text: sec.content || sec.text || '',
-          source: sec.name || sec.title || 'Секція',
-        });
-      }
-    }
-  }
-
-  // compare_documents
-  if (parsed.comparison || parsed.differences) {
-    const compText = parsed.comparison || parsed.summary || '';
-    if (compText) {
-      citations.push({
-        text: compText,
-        source: 'Порівняння документів',
-      });
-    }
-  }
-
-  return { decisions: [], citations, documents };
-}
-
-function extractRetrieveLegalSourcesEvidence(toolName: string, parsed: any): EvidenceResult {
-  if (toolName !== 'retrieve_legal_sources') {
-    return { decisions: [], citations: [], documents: [] };
-  }
-  const decisions: Decision[] = [];
-  const citations: Citation[] = [];
-
-  if (parsed.cases && Array.isArray(parsed.cases)) {
-    for (const c of parsed.cases) {
-      decisions.push({
-        id: `rls-${c.id || Math.random().toString(36).slice(2, 8)}`,
-        number: c.id || 'N/A',
-        court: c.court || '',
-        date: c.date || '',
-        summary: c.text?.slice(0, 300) || c.title || '',
-        relevance: 70,
-        status: 'active',
-      });
-    }
-  }
-  if (parsed.laws && Array.isArray(parsed.laws)) {
-    for (const l of parsed.laws) {
-      citations.push({
-        text: l.text || l.full_text || l.title || '',
-        source: l.article ? `${l.title || l.rada_id || ''}, ст. ${l.article}` : (l.title || 'Норма'),
-      });
-    }
-  }
-  return { decisions, citations, documents: [] };
-}
-
-function extractRadaEvidence(toolName: string, parsed: any): EvidenceResult {
-  const citations: Citation[] = [];
-  const documents: VaultDocument[] = [];
-  const radaTools = [
-    'rada_search_parliament_bills',
-    'rada_get_deputy_info',
-    'rada_search_legislation_text',
-    'rada_analyze_voting_record',
-  ];
-  if (!radaTools.some((t) => toolName === t)) {
-    return { decisions: [], citations, documents };
-  }
-
-  // Bills
-  if (parsed.bills && Array.isArray(parsed.bills)) {
-    for (const bill of parsed.bills) {
-      documents.push({
-        id: `bill-${bill.id || bill.number || Math.random().toString(36).slice(2, 8)}`,
-        title: bill.title || bill.name || `Законопроект ${bill.number || ''}`,
-        type: 'legislation',
-        metadata: {
-          snippet: bill.summary || bill.description || '',
-          status: bill.status,
-          number: bill.number,
-          date: bill.date || bill.registration_date,
-        },
-      });
-    }
-  }
-
-  // Deputy info (single)
-  if (parsed.name && (parsed.faction || parsed.party || parsed.deputy_id)) {
-    documents.push({
-      id: `deputy-${parsed.deputy_id || parsed.id || Math.random().toString(36).slice(2, 8)}`,
-      title: parsed.name || parsed.full_name || 'Народний депутат',
-      type: 'other',
-      metadata: {
-        snippet: [parsed.faction || parsed.party, parsed.region, parsed.position].filter(Boolean).join(' \u2022 '),
-        deputy_id: parsed.deputy_id || parsed.id,
-      },
-    });
-  }
-
-  // Deputies array
-  if (parsed.deputies && Array.isArray(parsed.deputies)) {
-    for (const dep of parsed.deputies) {
-      documents.push({
-        id: `deputy-${dep.id || Math.random().toString(36).slice(2, 8)}`,
-        title: dep.name || dep.full_name || 'Депутат',
-        type: 'other',
-        metadata: {
-          snippet: [dep.faction || dep.party, dep.region].filter(Boolean).join(' \u2022 '),
-        },
-      });
-    }
-  }
-
-  // Legislation text search
-  if (parsed.results && Array.isArray(parsed.results) && toolName === 'rada_search_legislation_text') {
-    for (const r of parsed.results) {
-      citations.push({
-        text: (r.text || r.snippet || r.content || '').slice(0, 500),
-        source: r.title || r.law_title || r.source || 'Законодавство',
-      });
-    }
-  }
-
-  // Voting record
-  if (parsed.votings && Array.isArray(parsed.votings)) {
-    for (const v of parsed.votings) {
-      citations.push({
-        text: `За: ${v.yes || 0}, Проти: ${v.no || 0}, Утримались: ${v.abstain || 0}${v.result ? ` \u2014 ${v.result}` : ''}`,
-        source: v.title || v.bill_title || 'Голосування',
-      });
-    }
-  }
-  if (parsed.voting_summary) {
-    citations.push({
-      text: typeof parsed.voting_summary === 'string'
-        ? parsed.voting_summary.slice(0, 500)
-        : JSON.stringify(parsed.voting_summary).slice(0, 500),
-      source: 'Аналіз голосувань',
-    });
-  }
-
-  return { decisions: [], citations, documents };
-}
-
-function extractRegistryEvidence(toolName: string, parsed: any): EvidenceResult {
-  const citations: Citation[] = [];
-  const documents: VaultDocument[] = [];
-  const registryTools = [
-    'openreyestr_search_entities',
-    'openreyestr_get_entity_details',
-    'openreyestr_search_beneficiaries',
-    'openreyestr_get_by_edrpou',
-    'openreyestr_get_statistics',
-    'openreyestr_search_enforcement_proceedings',
-    'openreyestr_search_debtors',
-    'openreyestr_search_bankruptcy_cases',
-    'openreyestr_search_notaries',
-    'openreyestr_search_court_experts',
-    'openreyestr_search_arbitration_managers',
-  ];
-  if (!registryTools.some((t) => toolName === t)) {
-    return { decisions: [], citations, documents };
-  }
-
-  // Entity search results
-  const entities = parsed.entities || parsed.results || [];
-  if (Array.isArray(entities)) {
-    for (const e of entities) {
-      documents.push({
-        id: `entity-${e.id || e.edrpou || Math.random().toString(36).slice(2, 8)}`,
-        title: e.name || e.full_name || e.short_name || 'Юрособа',
-        type: 'other',
-        metadata: {
-          snippet: [e.edrpou && `ЄДРПОУ: ${e.edrpou}`, e.address, e.status].filter(Boolean).join(' \u2022 '),
-          edrpou: e.edrpou,
-          status: e.status,
-        },
-      });
-    }
-  }
-
-  // Single entity details
-  if (parsed.name && parsed.edrpou && !parsed.entities) {
-    documents.push({
-      id: `entity-${parsed.edrpou}`,
-      title: parsed.name || parsed.full_name || 'Юрособа',
-      type: 'other',
-      metadata: {
-        snippet: [
-          `ЄДРПОУ: ${parsed.edrpou}`,
-          parsed.address,
-          parsed.status,
-          parsed.head && `Керівник: ${parsed.head}`,
-        ].filter(Boolean).join(' \u2022 '),
-        edrpou: parsed.edrpou,
-        status: parsed.status,
-      },
-    });
-  }
-
-  // Beneficiaries
-  if (parsed.beneficiaries && Array.isArray(parsed.beneficiaries)) {
-    for (const b of parsed.beneficiaries) {
-      documents.push({
-        id: `benef-${b.id || Math.random().toString(36).slice(2, 8)}`,
-        title: b.name || b.full_name || 'Бенефіціар',
-        type: 'other',
-        metadata: {
-          snippet: [b.share && `Частка: ${b.share}%`, b.country, b.entity_name].filter(Boolean).join(' \u2022 '),
-        },
-      });
-    }
-  }
-
-  // Statistics
-  if (parsed.statistics || parsed.total_count != null) {
-    citations.push({
-      text: typeof parsed.statistics === 'string'
-        ? parsed.statistics
-        : `Загалом: ${parsed.total_count || 0}. ${parsed.summary || ''}`.trim(),
-      source: 'Статистика реєстру',
-    });
-  }
-
-  // Enforcement proceedings
-  if (parsed.enforcement_proceedings || (parsed.results && Array.isArray(parsed.results) && parsed.results[0]?.proceeding_number)) {
-    const proceedings = parsed.enforcement_proceedings || parsed.results || [];
-    if (Array.isArray(proceedings)) {
-      for (const p of proceedings) {
-        documents.push({
-          id: `enforcement-${p.id || p.proceeding_number || Math.random().toString(36).slice(2, 8)}`,
-          title: `Виконавче провадження ${p.proceeding_number || ''}`,
-          type: 'other',
-          metadata: {
-            snippet: [
-              p.debtor_name && `Боржник: ${p.debtor_name}`,
-              p.creditor_name && `Стягувач: ${p.creditor_name}`,
-              p.proceeding_status && `Статус: ${p.proceeding_status}`,
-              p.enforcement_agency && `Виконавець: ${p.enforcement_agency}`,
-            ].filter(Boolean).join(' \u2022 '),
-            status: p.proceeding_status,
-          },
-        });
-      }
-    }
-  }
-
-  // Debtors
-  if (parsed.debtors || (parsed.results && Array.isArray(parsed.results) && parsed.results[0]?.debtor_name)) {
-    const debtors = parsed.debtors || parsed.results || [];
-    if (Array.isArray(debtors)) {
-      for (const d of debtors) {
-        documents.push({
-          id: `debtor-${d.id || d.edrpou || Math.random().toString(36).slice(2, 8)}`,
-          title: d.debtor_name || 'Боржник',
-          type: 'other',
-          metadata: {
-            snippet: [
-              d.edrpou && `ЄДРПОУ: ${d.edrpou}`,
-              d.total_debt && `Загальний борг: ${d.total_debt}`,
-              d.execution_proceedings && `Проваджень: ${d.execution_proceedings}`,
-            ].filter(Boolean).join(' \u2022 '),
-          },
-        });
-      }
-    }
-  }
-
-  // Bankruptcy cases
-  if (parsed.bankruptcy_cases || (parsed.results && Array.isArray(parsed.results) && parsed.results[0]?.case_number)) {
-    const cases = parsed.bankruptcy_cases || parsed.results || [];
-    if (Array.isArray(cases)) {
-      for (const c of cases) {
-        documents.push({
-          id: `bankruptcy-${c.id || c.case_number || Math.random().toString(36).slice(2, 8)}`,
-          title: `Банкрутство ${c.case_number || ''}`,
-          type: 'other',
-          metadata: {
-            snippet: [
-              c.debtor_name && `Боржник: ${c.debtor_name}`,
-              c.case_status && `Статус: ${c.case_status}`,
-              c.sanction && `Санкція: ${c.sanction}`,
-            ].filter(Boolean).join(' \u2022 '),
-          },
-        });
-      }
-    }
-  }
-
-  // Notaries
-  if (parsed.notaries || (parsed.results && Array.isArray(parsed.results) && parsed.results[0]?.notary_name)) {
-    const notaries = parsed.notaries || parsed.results || [];
-    if (Array.isArray(notaries)) {
-      for (const n of notaries) {
-        documents.push({
-          id: `notary-${n.id || n.notary_number || Math.random().toString(36).slice(2, 8)}`,
-          title: n.notary_name || 'Нотаріус',
-          type: 'other',
-          metadata: {
-            snippet: [
-              n.license_number && `Ліцензія: ${n.license_number}`,
-              n.address,
-              n.status,
-            ].filter(Boolean).join(' \u2022 '),
-          },
-        });
-      }
-    }
-  }
-
-  // Court experts
-  if (parsed.court_experts || (parsed.results && Array.isArray(parsed.results) && parsed.results[0]?.expert_name)) {
-    const experts = parsed.court_experts || parsed.results || [];
-    if (Array.isArray(experts)) {
-      for (const e of experts) {
-        documents.push({
-          id: `expert-${e.id || e.registration_number || Math.random().toString(36).slice(2, 8)}`,
-          title: e.expert_name || 'Судовий експерт',
-          type: 'other',
-          metadata: {
-            snippet: [
-              e.registration_number && `Реєстр №: ${e.registration_number}`,
-              e.specialization,
-              e.status,
-            ].filter(Boolean).join(' \u2022 '),
-          },
-        });
-      }
-    }
-  }
-
-  // Arbitration managers
-  if (parsed.arbitration_managers || (parsed.results && Array.isArray(parsed.results) && parsed.results[0]?.manager_name)) {
-    const managers = parsed.arbitration_managers || parsed.results || [];
-    if (Array.isArray(managers)) {
-      for (const m of managers) {
-        documents.push({
-          id: `manager-${m.id || m.registration_number || Math.random().toString(36).slice(2, 8)}`,
-          title: m.manager_name || 'Арбітражний керуючий',
-          type: 'other',
-          metadata: {
-            snippet: [
-              m.registration_number && `Реєстр №: ${m.registration_number}`,
-              m.address,
-              m.status,
-            ].filter(Boolean).join(' \u2022 '),
-          },
-        });
-      }
-    }
-  }
-
-  return { decisions: [], citations, documents };
-}
-
-function extractProceduralEvidence(toolName: string, parsed: any): EvidenceResult {
-  const citations: Citation[] = [];
-  const proceduralTools = [
-    'calculate_procedural_deadlines',
-    'build_procedural_checklist',
-    'calculate_monetary_claims',
-  ];
-  if (!proceduralTools.some((t) => toolName === t)) {
-    return { decisions: [], citations, documents: [] };
-  }
-
-  // Deadlines
-  if (parsed.deadlines && Array.isArray(parsed.deadlines)) {
-    for (const dl of parsed.deadlines) {
-      citations.push({
-        text: `${dl.description || dl.action || dl.name}: ${dl.deadline || dl.date || dl.days_left ? `${dl.days_left} днів` : ''}`,
-        source: dl.legal_basis || dl.norm || 'Процесуальний строк',
-      });
-    }
-  }
-
-  // Checklist
-  if (parsed.checklist && Array.isArray(parsed.checklist)) {
-    for (const item of parsed.checklist) {
-      citations.push({
-        text: `${item.step || item.action || item.description || ''}${item.deadline ? ` (до ${item.deadline})` : ''}`,
-        source: item.legal_basis || item.norm || 'Процесуальний чеклист',
-      });
-    }
-  }
-  if (parsed.items && Array.isArray(parsed.items)) {
-    for (const item of parsed.items) {
-      citations.push({
-        text: item.description || item.text || item.name || '',
-        source: item.legal_basis || 'Чеклист',
-      });
-    }
-  }
-
-  // Monetary claims
-  if (parsed.total != null || parsed.amount != null || parsed.calculation) {
-    citations.push({
-      text: parsed.calculation
-        || `Загальна сума: ${parsed.total || parsed.amount || 0} грн${parsed.breakdown ? `. ${parsed.breakdown}` : ''}`,
-      source: parsed.legal_basis || 'Розрахунок грошових вимог',
-    });
-  }
-  if (parsed.components && Array.isArray(parsed.components)) {
-    for (const comp of parsed.components) {
-      citations.push({
-        text: `${comp.name || comp.type || 'Складова'}: ${comp.amount || 0} грн`,
-        source: comp.legal_basis || 'Складова вимоги',
-      });
-    }
-  }
-
-  return { decisions: [], citations, documents: [] };
-}
-
-function extractDDEvidence(toolName: string, parsed: any): EvidenceResult {
-  const citations: Citation[] = [];
-  const ddTools = [
-    'generate_dd_report',
-    'risk_scoring',
-    'format_answer_pack',
-  ];
-  if (!ddTools.some((t) => toolName === t)) {
-    return { decisions: [], citations, documents: [] };
-  }
-
-  // DD report
-  if (parsed.report || parsed.summary || parsed.findings) {
-    const reportText = parsed.report || parsed.summary || '';
-    if (reportText) {
-      citations.push({
-        text: (typeof reportText === 'string' ? reportText : JSON.stringify(reportText)).slice(0, 500),
-        source: 'Due Diligence звіт',
-      });
-    }
-    if (parsed.findings && Array.isArray(parsed.findings)) {
-      for (const f of parsed.findings) {
-        citations.push({
-          text: (f.description || f.text || f.finding || '').slice(0, 500),
-          source: f.category || f.type || 'Висновок DD',
-        });
-      }
-    }
-  }
-
-  // Risk scoring
-  if (parsed.risk_score != null || parsed.score != null) {
-    citations.push({
-      text: `Рівень ризику: ${parsed.risk_score || parsed.score}${parsed.risk_level ? ` (${parsed.risk_level})` : ''}. ${parsed.explanation || parsed.summary || ''}`.trim(),
-      source: 'Скоринг ризиків',
-    });
-  }
-  if (parsed.risks && Array.isArray(parsed.risks)) {
-    for (const r of parsed.risks) {
-      citations.push({
-        text: `${r.name || r.type || 'Ризик'}: ${r.score || r.level || ''}. ${r.description || ''}`.trim(),
-        source: r.category || 'Ризик',
-      });
-    }
-  }
-
-  // Answer pack
-  if (parsed.answers && Array.isArray(parsed.answers)) {
-    for (const a of parsed.answers) {
-      citations.push({
-        text: (a.answer || a.text || a.content || '').slice(0, 500),
-        source: a.question || a.topic || 'Відповідь',
-      });
-    }
-  }
-
-  return { decisions: [], citations, documents: [] };
-}
-
-// ---- Main entry point ----
-
-function mergeEvidence(target: EvidenceResult, source: EvidenceResult): void {
-  target.decisions.push(...source.decisions);
-  target.citations.push(...source.citations);
-  target.documents.push(...source.documents);
-}
-
-/**
- * Extract Decision[], Citation[], VaultDocument[] from a tool_result SSE event.
- * Handles all court-case, legislation, RADA, registry, vault and procedural tools.
- */
-export function extractEvidenceFromToolResult(
-  toolName: string,
-  rawResult: any
-): EvidenceResult {
-  const result: EvidenceResult = { decisions: [], citations: [], documents: [] };
-
-  const parsed = parseToolResultContent(rawResult);
-  if (!parsed) return result;
-
-  mergeEvidence(result, extractCourtEvidence(toolName, parsed));
-  mergeEvidence(result, extractLegislationEvidence(toolName, parsed));
-  mergeEvidence(result, extractProceduralNormEvidence(toolName, rawResult));
-  mergeEvidence(result, extractVaultEvidence(toolName, parsed));
-  mergeEvidence(result, extractRetrieveLegalSourcesEvidence(toolName, parsed));
-  mergeEvidence(result, extractRadaEvidence(toolName, parsed));
-  mergeEvidence(result, extractRegistryEvidence(toolName, parsed));
-  mergeEvidence(result, extractProceduralEvidence(toolName, parsed));
-  mergeEvidence(result, extractDDEvidence(toolName, parsed));
-
-  return result;
-}
+export { extractEvidenceFromToolResult, classifyDocumentType } from './evidence/index';
+export type { EvidenceResult } from './evidence/index';
