+++ lexwebapp/src/hooks/chat/evidence/court.ts
@@ -1,0 +1,111 @@
+import type { Decision } from '../../../types/models/Message';
+import type { EvidenceResult } from './types';
+import { classifyDocumentType, courtDocUrl } from './parse';
+
+const COURT_TOOLS = [
+  'search_legal_precedents',
+  'search_supreme_court_practice',
+  'get_case_documents_chain',
+  'find_similar_fact_pattern_cases',
+  'compare_practice_pro_contra',
+  'get_court_decision',
+  'count_cases_by_party',
+];
+
+export function extractCourtEvidence(toolName: string, parsed: any): EvidenceResult {
+  const decisions: Decision[] = [];
+  if (!COURT_TOOLS.some((t) => toolName.includes(t) || toolName === t)) {
+    return { decisions, citations: [], documents: [] };
+  }
+
+  // source_case (single)
+  if (parsed.source_case) {
+    const sc = parsed.source_case;
+    decisions.push({
+      id: `sc-${sc.doc_id || Date.now()}`,
+      number: sc.cause_num || sc.case_number || 'N/A',
+      court: sc.court_code || sc.court || '',
+      date: sc.adjudication_date || sc.date || '',
+      summary: sc.title || sc.resolution || '',
+      relevance: 100,
+      status: 'active',
+      documentType: classifyDocumentType(sc),
+      externalUrl: courtDocUrl(sc.doc_id),
+    });
+  }
+
+  // similar_cases / results array
+  const cases = parsed.similar_cases || parsed.results || parsed.cases || parsed.precedents || [];
+  for (const c of cases) {
+    decisions.push({
+      id: `d-${c.doc_id || c.id || Math.random().toString(36).slice(2, 8)}`,
+      number: c.cause_num || c.case_number || c.number || 'N/A',
+      court: c.court_code || c.court || '',
+      date: c.adjudication_date || c.date || '',
+      summary: c.title || c.resolution || c.summary || c.similarity_reason
+        || (Array.isArray(c.snippets) ? c.snippets.join(' ') : '') || '',
+      relevance: c.similarity
+        ? Math.round(c.similarity * 100)
+        : c.relevance
+          ? Math.round(c.relevance * 100)
+          : 70,
+      status: 'active',
+      documentType: classifyDocumentType(c),
+      externalUrl: courtDocUrl(c.doc_id),
+    });
+  }
+
+  // get_case_documents_chain format
+  let chainDocs: any[] = [];
+  if (parsed.documents && Array.isArray(parsed.documents)) {
+    chainDocs = parsed.documents;
+  } else if (parsed.grouped_documents && typeof parsed.grouped_documents === 'object') {
+    chainDocs = Object.values(parsed.grouped_documents).flat();
+  }
+  for (const doc of chainDocs) {
+    decisions.push({
+      id: `chain-${doc.doc_id || Math.random().toString(36).slice(2, 8)}`,
+      number: doc.case_number || parsed.case_number || doc.title || 'N/A',
+      court: doc.court || doc.instance || '',
+      date: doc.date || '',
+      summary: doc.resolution || doc.title || '',
+      relevance: 80,
+      status: 'active',
+      documentType: classifyDocumentType(doc),
+      externalUrl: courtDocUrl(doc.doc_id),
+    });
+  }
+
+  // compare_practice_pro_contra format
+  const proContraCases = [...(parsed.pro || []), ...(parsed.contra || [])];
+  for (const c of proContraCases) {
+    decisions.push({
+      id: `pc-${c.doc_id || Math.random().toString(36).slice(2, 8)}`,
+      number: c.case_number || 'N/A',
+      court: c.court || c.chamber || '',
+      date: c.date || '',
+      summary: c.snippet || '',
+      relevance: 70,
+      status: 'active',
+      documentType: classifyDocumentType(c),
+      externalUrl: courtDocUrl(c.doc_id),
+    });
+  }
+
+  // get_court_decision — single decision with sections
+  if (parsed.sections && Array.isArray(parsed.sections) && (parsed.doc_id || parsed.case_number)) {
+    const summarySection = parsed.sections.find((s: any) => s.type === 'DECISION' || s.type === 'COURT_REASONING');
+    decisions.push({
+      id: `gcd-${parsed.doc_id || Date.now()}`,
+      number: parsed.case_number || String(parsed.doc_id) || 'N/A',
+      court: '',
+      date: '',
+      summary: summarySection?.text?.slice(0, 300) || '',
+      relevance: 100,
+      status: 'active',
+      externalUrl: courtDocUrl(parsed.doc_id),
+    });
+  }
+
+  return { decisions, citations: [], documents: [] };
+}
